[TM-1105] Finished refactoring the Site PD page.
